WIP refactor facets
