+++ lists-ui/package.json
@@ -27,6 +27,7 @@
     "@mantine/modals": "^7.12.1",
     "@mantine/notifications": "^7.12.1",
     "@mantine/nprogress": "^7.11.2",
+    "@tabler/icons-react": "^3.31.0",
     "date-fns": "^4.1.0",
     "jwt-decode": "^4.0.0",
     "lodash-es": "^4.17.21",
