--- lists-ui/src/components/FiltersDrawer.module.css
@@ -1,22 +1,0 @@
-.filter {
-  padding: 8px 16px 8px 16px;
-
-  &:hover {
-    @mixin light {
-      background-color: var(--mantine-color-grey-1);
-    }
-    @mixin dark {
-      background-color: var(--mantine-color-dark-5);
-    }
-  }
-}
-
-.heading {
-  position: sticky;
-  top: 64px;
-  padding-left: 16px;
-  padding-right: 16px;
-  padding-bottom: 10px;
-  border-radius: 0;
-  z-index: 100;
-}
