+++ lists-ui/src/locale/en.json
@@ -1,14 +1,14 @@
 {
-  "CONSERVATION_LIST": "Conservation List",
-  "SENSITIVE_LIST": "Sensitive List",
-  "COMMON_TRAIT": "Traits List",
-  "LOCAL_LIST": "Local List",
-  "SPECIES_CHARACTERS": "Character List",
-  "INVASIVE": "Invasive species List",
-  "COMMON_HABITAT": "Common Habitat List",
-  "PROFILE": "Profile List",
-  "TEST": "Testing List",
-  "OTHER": "Other List",
+  "CONSERVATION_LIST": "Conservation list",
+  "SENSITIVE_LIST": "Sensitive list",
+  "COMMON_TRAIT": "Common traits list",
+  "LOCAL_LIST": "Local list",
+  "SPECIES_CHARACTERS": "Speceis characters list",
+  "INVASIVE": "Invasive species list",
+  "COMMON_HABITAT": "Common habitat list",
+  "PROFILE": "Profile list",
+  "TEST": "Testing list",
+  "OTHER": "Other list",
 
   "classification.kingdom": "Kingdom",
   "classification.phylum": "Phylum",
@@ -23,7 +23,7 @@
   "isAuthoritative": "Authoritative list",
   "listType": "List type",
   "isBIE": "Visible on species pages",
-  "isSDS": "Used by sensitive data service",
+  "isSDS": "Sensitive data service",
   "tags": "Tags",
   "hasRegion": "Region specified",
   "true": "yes",
@@ -43,6 +43,7 @@
   "list.isAuthoritative": "Authoritative",
   "list.isSDS": "Sensitive data service",
   "list.isBIE": "Species pages",
+  "list.properties": "List properties",
 
   "scientificName": "Scientific name",
   "scientificNameAuthority": "Scientific name authority",
